Revert "Make src/crypto and src/controller compile with -Wconversion (#2810)"
This reverts commit 3ec5a53081fc874641d4635ab43faa1b827e21ab.

The commit included extraneous changes that should not have landed as part of that PR.

+++ src/transport/BLE.cpp
@@ -138,7 +138,7 @@ CHIP_ERROR BLE::SendMessage(const PacketHeader & header, Header::Flags payloadFl
 {
     CHIP_ERROR err          = CHIP_NO_ERROR;
     const size_t headerSize = header.EncodeSizeBytes();
-    uint16_t actualEncodedHeaderSize;
+    size_t actualEncodedHeaderSize;
 
     VerifyOrExit(address.GetTransportType() == Type::kBle, err = CHIP_ERROR_INVALID_ARGUMENT);
     VerifyOrExit(mState == State::kInitialized, err = CHIP_ERROR_INCORRECT_STATE);
